% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccumulatR-package.R
\docType{package}
\name{AccumulatR-package}
\alias{AccumulatR}
\alias{AccumulatR-package}
\title{AccumulatR: Simulate and fit evidence-accumulation models}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}
}
\details{
\code{AccumulatR} is for researchers who work with response-time models in which
behavior is generated by competing or interacting accumulators. The package
lets you describe a model in psychologically meaningful terms, simulate
behavioral data, and evaluate likelihoods for fitting.

The main workflow is:
\enumerate{
\item Start with \code{\link[=race_spec]{race_spec()}}.
\item Add accumulators, pools, outcomes, and timing relations.
\item Call \code{\link[=finalize_model]{finalize_model()}} and supply values with \code{\link[=build_param_matrix]{build_param_matrix()}}.
\item Generate observations with \code{\link[=simulate]{simulate()}}, or use experimental data.
\item Call \code{\link[=prepare_data]{prepare_data()}} and \code{\link[=make_context]{make_context()}}, then evaluate candidate
parameters with \code{\link[=log_likelihood]{log_likelihood()}}.
}

See \code{vignette("getting_started", package = "AccumulatR")} for a worked example.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/niekstevenson/AccumulatR}
  \item \url{https://niekstevenson.github.io/AccumulatR/}
  \item Report bugs at \url{https://github.com/niekstevenson/AccumulatR/issues}
}

}
\author{
\strong{Maintainer}: Niek Stevenson \email{niek.stevenson@gmail.com}

}
\keyword{package}
