% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_definition.R
\name{add_group}
\alias{add_group}
\title{Add a group of members}
\usage{
add_group(spec, id, members, attrs = list())
}
\arguments{
\item{spec}{race_spec object}

\item{id}{Group id}

\item{members}{Member ids}

\item{attrs}{Attributes list}
}
\value{
Updated race_spec
}
\description{
Add a group of members
}
\examples{
spec <- race_spec()
spec <- add_group(spec, "G1", members = c("A", "B"),
  attrs = list(shared_params = list(q = 0.2)))
}
