% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_definition.R
\name{all_of}
\alias{all_of}
\title{Define a response that requires several processes to finish}
\usage{
all_of(...)
}
\arguments{
\item{...}{Accumulator labels, pool labels, or response expressions.}
}
\value{
An expression object.
}
\description{
Return a rule that finishes when all required events have completed.
Any \code{\link[=none_of]{none_of()}} conditions are checked at that finishing time.
}
\examples{
all_of("A", "B")
}
