% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_definition.R
\name{build_model}
\alias{build_model}
\title{Finalize a model (deprecated alias)}
\usage{
build_model(spec)
}
\arguments{
\item{spec}{race_spec object}
}
\value{
model_structure
}
\description{
Finalize a model (deprecated alias)
}
\examples{
spec <- race_spec()
spec <- add_accumulator(spec, "A", "lognormal",
  params = list(meanlog = 0, sdlog = 0.1))
spec <- add_outcome(spec, "A_win", "A")
build_model(spec)
}
