% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_definition.R
\name{component}
\alias{component}
\title{Define a component entry}
\usage{
component(id, weight = NULL, attrs = list())
}
\arguments{
\item{id}{Component id}

\item{weight}{Optional weight}

\item{attrs}{Optional attributes list}
}
\value{
Component list
}
\description{
Define a component entry
}
\examples{
component("fast", weight = 0.7)
}
