% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_definition.R
\name{expr_guard}
\alias{expr_guard}
\title{Build a blocking rule explicitly}
\usage{
expr_guard(blocker, reference)
}
\arguments{
\item{blocker}{Blocking expression or accumulator label.}

\item{reference}{Target expression or accumulator label.}
}
\value{
A guarded expression object.
}
\description{
Most users will prefer `inhibit()`, but `expr_guard()` is available when you
want to construct the guarded expression directly.
}
\examples{
expr_guard("B", "A")
}
