% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_definition.R
\name{first_of}
\alias{first_of}
\title{Define a response that occurs when the first listed process finishes}
\usage{
first_of(...)
}
\arguments{
\item{...}{Accumulator labels, pool labels, or response expressions.}
}
\value{
An expression object.
}
\description{
Return a rule whose finishing time is the earliest completion among its
arguments. Each argument must contain an event that can generate a response.
}
\examples{
first_of("A", "B")
}
