% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_definition.R
\name{par_names}
\alias{par_names}
\title{List the free parameters implied by a model}
\usage{
par_names(model)
}
\arguments{
\item{model}{A `race_spec` or finalized `model_structure` object.}
}
\value{
A character vector of parameter names.
}
\description{
List the free parameters implied by a model
}
\examples{
spec <- race_spec()
spec <- add_accumulator(spec, "A", "lognormal")
spec <- add_outcome(spec, "A_win", "A")
par_names(spec)
}
