% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_definition.R
\name{sampled_pars}
\alias{sampled_pars}
\title{List the free parameters implied by a model}
\usage{
sampled_pars(model)
}
\arguments{
\item{model}{A `race_spec` or related model object.}
}
\value{
A character vector of parameter names.
}
\description{
List the free parameters implied by a model
}
\examples{
spec <- race_spec()
spec <- add_accumulator(spec, "A", "lognormal")
spec <- add_outcome(spec, "A_win", "A")
sampled_pars(spec)
}
