% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_definition.R
\name{set_mixture_options}
\alias{set_mixture_options}
\title{Control how mixture components are combined}
\usage{
set_mixture_options(spec, mode = "sample", reference = NULL)
}
\arguments{
\item{spec}{A `race_spec` object.}

\item{mode}{Mixture mode.}

\item{reference}{Optional reference component.}
}
\value{
The updated `race_spec`.
}
\description{
Control how mixture components are combined
}
