% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_definition.R
\name{shared_q}
\alias{shared_q}
\title{Shared trigger probability placeholder}
\usage{
shared_q()
}
\value{
Placeholder used in model specs
}
\description{
Shared trigger probability placeholder
}
\examples{
shared_q()
}
