% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_definition.R
\name{add_accumulator}
\alias{add_accumulator}
\title{Add an accumulator to a model}
\usage{
add_accumulator(spec, id, dist, onset = 0)
}
\arguments{
\item{spec}{A `race_spec` object.}

\item{id}{Label for the accumulator.}

\item{dist}{Distribution family used for that accumulator.}

\item{onset}{Start time for the accumulator. This can be a fixed numeric
onset or a chained onset created with `after()`.}
}
\value{
The updated `race_spec`.
}
\description{
Add an accumulator to a model
}
\examples{
spec <- race_spec()
spec <- add_accumulator(spec, "A", "lognormal")
}
