% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_definition.R
\name{add_accumulator}
\alias{add_accumulator}
\title{Add an accumulator definition}
\usage{
add_accumulator(
  spec,
  id,
  dist,
  onset = 0,
  q = 0,
  tags = list(),
  params = NULL,
  ...
)
}
\arguments{
\item{spec}{race_spec object}

\item{id}{Accumulator id}

\item{dist}{Distribution name}

\item{onset}{Onset shift}

\item{q}{Guess probability (0-1)}

\item{tags}{Optional tags list}

\item{params}{Optional parameter list}

\item{...}{Additional named parameters}
}
\value{
Updated race_spec
}
\description{
Add an accumulator definition
}
\examples{
spec <- race_spec()
spec <- add_accumulator(spec, "A", "lognormal",
  params = list(meanlog = 0, sdlog = 0.1)
)
}
