% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_definition.R
\name{add_component}
\alias{add_component}
\title{Add a group of members}
\usage{
add_component(
  spec,
  id,
  members,
  weight = NULL,
  weight_param = NULL,
  attrs = list()
)
}
\arguments{
\item{spec}{race_spec object}

\item{id}{Component id (e.g. "fast")}

\item{members}{Character vector of accumulator IDs belonging to this component.}

\item{weight}{Optional fixed weight (0-1).}

\item{weight_param}{Optional parameter name for weight.}

\item{attrs}{Additional attributes.}
}
\value{
Updated race_spec
}
\description{
Add a group of members
}
