% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_definition.R
\name{add_outcome}
\alias{add_outcome}
\title{Add an outcome definition}
\usage{
add_outcome(spec, label, expr, options = list())
}
\arguments{
\item{spec}{race_spec object}

\item{label}{Outcome label}

\item{expr}{Outcome expression}

\item{options}{Optional list of options}
}
\value{
Updated race_spec
}
\description{
Add an outcome definition
}
\examples{
spec <- race_spec()
spec <- add_outcome(spec, "A_win", "A")
}
