% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_definition.R
\name{add_pool}
\alias{add_pool}
\title{Add a pool definition}
\usage{
add_pool(spec, id, members, k = 1L)
}
\arguments{
\item{spec}{race_spec object}

\item{id}{Pool id}

\item{members}{Member ids}

\item{k}{Threshold parameter}
}
\value{
Updated race_spec
}
\description{
Add a pool definition
}
\examples{
spec <- race_spec()
spec <- add_pool(spec, "P1", members = c("A", "B"), k = 1L)
}
