% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_definition.R
\name{all_of}
\alias{all_of}
\title{Define a response that requires several processes to finish}
\usage{
all_of(...)
}
\arguments{
\item{...}{Accumulator labels or expression objects to combine with AND.}
}
\value{
An expression object.
}
\description{
Define a response that requires several processes to finish
}
\examples{
all_of("A", "B")
}
