% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_definition.R
\name{all_of}
\alias{all_of}
\title{All-of expression helper}
\usage{
all_of(...)
}
\arguments{
\item{...}{Expressions or labels to combine with AND}
}
\value{
Expression list
}
\description{
All-of expression helper
}
\examples{
all_of("A", "B")
}
