% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_definition.R
\name{build_outcome_expr}
\alias{build_outcome_expr}
\title{Normalize an outcome expression}
\usage{
build_outcome_expr(expr)
}
\arguments{
\item{expr}{Expression or symbol describing an event/guard}
}
\value{
Expression list used in model specs
}
\description{
Normalize an outcome expression
}
\examples{
build_outcome_expr(quote(A & !B))
}
