% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_definition.R
\name{expr_guard}
\alias{expr_guard}
\title{Build an expression guard node}
\usage{
expr_guard(blocker, reference, unless = list())
}
\arguments{
\item{blocker}{Blocker expression or id}

\item{reference}{Reference expression or id}

\item{unless}{Optional list of unless expressions}
}
\value{
Guard expression list
}
\description{
Build an expression guard node
}
\examples{
expr_guard("B", "A")
}
