% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator_new.R
\name{finalize_model}
\alias{finalize_model}
\alias{build_model_structure}
\title{Finalize a model for simulation/likelihood}
\usage{
finalize_model(model)

build_model_structure(model)
}
\arguments{
\item{model}{Model specification}
}
\value{
A model_structure list
}
\description{
Finalize a model for simulation/likelihood
}
\note{
Deprecated: use finalize_model().
}
\examples{
spec <- race_spec()
spec <- add_accumulator(spec, "A", "lognormal",
  params = list(meanlog = 0, sdlog = 0.1))
spec <- add_outcome(spec, "A_win", "A")
finalize_model(spec)
}
