% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_definition.R
\name{finalize_model}
\alias{finalize_model}
\title{Finalize a model for simulation/likelihood}
\usage{
finalize_model(model)
}
\arguments{
\item{model}{Model specification}
}
\value{
A model_structure list
}
\description{
Finalize a model for simulation/likelihood
}
\examples{
spec <- race_spec()
spec <- add_accumulator(spec, "A", "lognormal")
spec <- add_outcome(spec, "A_win", "A")
finalize_model(spec)
}
