% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_definition.R
\name{first_of}
\alias{first_of}
\title{Define a response that occurs when the first listed process finishes}
\usage{
first_of(...)
}
\arguments{
\item{...}{Accumulator labels or expression objects to combine with OR.}
}
\value{
An expression object.
}
\description{
Define a response that occurs when the first listed process finishes
}
\examples{
first_of("A", "B")
}
