% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_definition.R
\name{first_of}
\alias{first_of}
\title{First-of expression helper}
\usage{
first_of(...)
}
\arguments{
\item{...}{Expressions or labels to combine with OR}
}
\value{
Expression list
}
\description{
First-of expression helper
}
\examples{
first_of("A", "B")
}
