% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_definition.R
\name{inhibit}
\alias{inhibit}
\title{Define a response that is blocked by another process}
\usage{
inhibit(reference, by)
}
\arguments{
\item{reference}{Response rule or accumulator label to be blocked.}

\item{by}{Blocking process or expression.}
}
\value{
A guarded expression object.
}
\description{
Define a response that is blocked by another process
}
\examples{
inhibit("A", "B")
}
