% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_definition.R
\name{inhibit}
\alias{inhibit}
\title{Define an inhibitory relationship}
\usage{
inhibit(reference, by)
}
\arguments{
\item{reference}{Outcome expression or label to inhibit}

\item{by}{Blocking expression or label}
}
\value{
Guard expression list
}
\description{
Define an inhibitory relationship
}
\examples{
inhibit("A", "B")
}
