% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_definition.R
\name{none_of}
\alias{none_of}
\title{Define the absence of an event}
\usage{
none_of(expr)
}
\arguments{
\item{expr}{Accumulator label or expression to negate.}
}
\value{
An expression object.
}
\description{
Define the absence of an event
}
\examples{
none_of("A")
}
