% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_definition.R
\name{none_of}
\alias{none_of}
\alias{exclude}
\title{None-of expression helper}
\usage{
none_of(expr)

exclude(expr)
}
\arguments{
\item{expr}{Expression or label to negate}
}
\value{
Expression list
}
\description{
None-of expression helper
}
\examples{
none_of("A")
}
