% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_definition.R
\name{race_spec}
\alias{race_spec}
\title{Create an empty race specification}
\usage{
race_spec(n_outcomes = 1L)
}
\arguments{
\item{n_outcomes}{Number of observed outcomes to retain per trial.}
}
\value{
A race_spec object
}
\description{
Create an empty race specification
}
\examples{
race_spec()
}
