% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_param_interface.R
\name{response_probabilities}
\alias{response_probabilities}
\alias{response_probabilities.model_structure}
\alias{response_probabilities.default}
\title{Evaluate marginal response probabilities}
\usage{
response_probabilities(structure, params_df, include_na = TRUE, ...)

\method{response_probabilities}{model_structure}(structure, params_df, include_na = TRUE, ...)

\method{response_probabilities}{default}(structure, params_df, include_na = TRUE, ...)
}
\arguments{
\item{structure}{Finalized model structure.}

\item{params_df}{A parameter data frame or rectangular parameter matrix.}

\item{include_na}{If `TRUE`, include residual mass as `"NA"`.}

\item{...}{Unused; for S3 compatibility.}
}
\description{
Evaluate marginal response probabilities
}
