% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_definition.R
\name{sampled_pars}
\alias{sampled_pars}
\title{List sampled parameter names for a model}
\usage{
sampled_pars(model)
}
\arguments{
\item{model}{race_spec or race_model_spec}
}
\value{
Character vector of parameter names
}
\description{
List sampled parameter names for a model
}
\examples{
spec <- race_spec()
spec <- add_accumulator(spec, "A", "lognormal")
spec <- add_outcome(spec, "A_win", "A")
sampled_pars(spec)
}
