% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_definition.R
\name{set_metadata}
\alias{set_metadata}
\title{Set metadata on a specification}
\usage{
set_metadata(spec, ...)
}
\arguments{
\item{spec}{race_spec object}

\item{...}{Named metadata entries}
}
\value{
Updated race_spec
}
\description{
Set metadata on a specification
}
\examples{
spec <- race_spec()
spec <- set_metadata(spec, rel_tol = 1e-4)
}
