% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_definition.R
\name{set_mixture_options}
\alias{set_mixture_options}
\title{Set mixture options (global)}
\usage{
set_mixture_options(spec, mode = "sample", reference = NULL)
}
\arguments{
\item{spec}{race_spec object}

\item{mode}{Mixture mode (e.g. "sample")}

\item{reference}{Component ID to use as reference (base)}
}
\value{
Updated race_spec
}
\description{
Set mixture options (global)
}
